test_cli_*: move files into test/cli/*
