+++ setup.py
@@ -261,6 +261,7 @@ setup(name=APPNAME,
                 'allmydata.storage',
                 'allmydata.test',
                 'allmydata.test.mutable',
+                'allmydata.test.cli',
                 'allmydata.util',
                 'allmydata.web',
                 'allmydata.windows',
