+++ src/allmydata/test/cli/test_backup.py
@@ -12,8 +12,8 @@ from allmydata.util.fileutil import abspath_expanduser_unicode
 from allmydata.util.encodingutil import get_io_encoding, unicode_to_argv
 from allmydata.util.namespace import Namespace
 from allmydata.scripts import cli, backupdb
-from .common_util import StallMixin
-from .no_network import GridTestMixin
+from ..common_util import StallMixin
+from ..no_network import GridTestMixin
 from .test_cli import CLITestMixin, parse_options
 
 timeout = 480 # deep_check takes 360s on Zandr's linksys box, others take > 240s
