+++ src/allmydata/test/cli/test_magic_folder.py
@@ -9,7 +9,7 @@ from twisted.python import usage
 from allmydata.util.assertutil import precondition
 from allmydata.util import fileutil
 from allmydata.scripts.common import get_aliases
-from allmydata.test.no_network import GridTestMixin
+from ..no_network import GridTestMixin
 from .test_cli import CLITestMixin
 from allmydata.scripts import magic_folder_cli
 from allmydata.util.fileutil import abspath_expanduser_unicode
